+++ mithril-aggregator/src/commands/tools_command.rs
@@ -75,7 +75,6 @@ impl RecomputeCertificatesHashCommand {
             .with_context(|| "recompute-certificates-hash: database migration error")?;
 
         vacuum_database(&connection)
-            .await
             .with_context(|| "recompute-certificates-hash: database vacuum error")?;
 
         Ok(())