+++ mithril-aggregator/src/dependency_injection/builder.rs
@@ -281,7 +281,9 @@ impl DependenciesBuilder {
         &self,
         sqlite_file_name: &str,
         migrations: Vec<SqlMigration>,
+        do_vacuum_database: bool,
     ) -> Result<SqliteConnection> {
+        let logger = self.get_logger()?;
         let connection_builder = match self.configuration.environment {
             ExecutionEnvironment::Production => ConnectionBuilder::open_file(
                 &self.configuration.get_sqlite_dir().join(sqlite_file_name),
@@ -297,13 +299,22 @@ impl DependenciesBuilder {
             ),
         };
 
-        let connection = connection_builder
-            .with_node_type(ApplicationNodeType::Aggregator)
-            .with_options(&[
+        let connection_options = {
+            let mut options = vec![
                 ConnectionOptions::EnableForeignKeys,
                 ConnectionOptions::EnableWriteAheadLog,
-            ])
-            .with_logger(self.get_logger()?)
+            ];
+            if do_vacuum_database {
+                options.push(ConnectionOptions::Vacuum);
+            }
+
+            options
+        };
+
+        let connection = connection_builder
+            .with_node_type(ApplicationNodeType::Aggregator)
+            .with_options(&connection_options)
+            .with_logger(logger.clone())
             .with_migrations(migrations)
             .build()
             .map_err(|e| DependenciesBuilderError::Initialization {
@@ -332,6 +343,7 @@ impl DependenciesBuilder {
             self.sqlite_connection = Some(Arc::new(self.build_sqlite_connection(
                 SQLITE_FILE,
                 crate::database::migration::get_migrations(),
+                true,
             )?));
         }
 
@@ -349,10 +361,12 @@ impl DependenciesBuilder {
         let _connection = self.build_sqlite_connection(
             SQLITE_FILE_CARDANO_TRANSACTION,
             mithril_persistence::database::cardano_transaction_migration::get_migrations(),
+            // Don't vacuum the Cardano transactions database as it can be very large
+            false,
         )?;
 
         let connection_pool = Arc::new(SqliteConnectionPool::build(connection_pool_size, || {
-            self.build_sqlite_connection(SQLITE_FILE_CARDANO_TRANSACTION, vec![])
+            self.build_sqlite_connection(SQLITE_FILE_CARDANO_TRANSACTION, vec![], false)
                 .with_context(|| {
                     "Dependencies Builder can not build SQLite connection for Cardano transactions"
                 })